moved CustomTooltip to external file.
