+++ app/javascript/bundles/app/RewiewSatisfaction/components/CustomTooltip.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const CustomTooltip = ({ active, payload, label }) => {
+  if (active) {
+    return (
+      <div className="custom-tooltip">
+        <p className="label">{`${label} : ${payload[0].value}`}</p>
+      </div>
+    );
+  }
+
+  return null;
+};
+
+export default CustomTooltip
